+++ app/svads3d/screen.py
@@ -117,7 +117,7 @@ class Screen:
         self.eillposid_overlapmesh = []
         self.eillposid_nonoverlapmesh = []
 
-        self.optimize()
+        # self.optimize()
         self.meshing()
         self.compute_uv()
 
@@ -308,26 +308,26 @@ class Screen:
             for i in range(len(frag[1]))[2:]:
                 gmsh.model.occ.remove(frag[1][i], recursive=True)
 
-            ground = [[[14], [5, 0]],
+            ground = [[[14], [5, 0, 34, 38]],
                       [[16], [0]],
                       [[18], [1]],
-                      [[21], [1, 2]],
+                      [[21], [1, 2, 45, 49]],
                       [[20], [2]],
-                      [[19], [2, 3]],
+                      [[19], [2, 3, 47, 44]],
                       [[17], [3]],
                       [[15], [4]],
-                      [[13], [4, 5]],
+                      [[13], [4, 5, 37, 36]],
                       [[12], [5]]]
 
-            eillposid = [[[4], [5, 0]],
+            eillposid = [[[4], [5, 0, 3, 12]],
                          [[6], [0]],
                          [[8], [1]],
-                         [[10], [1, 2]],
+                         [[10], [1, 2, 24, 30]],
                          [[11], [2]],
-                         [[9], [2, 3]],
+                         [[9], [2, 3, 27, 21]],
                          [[7], [3]],
                          [[5], [4]],
-                         [[3], [4, 5]],
+                         [[3], [4, 5, 9, 6]],
                          [[2, 1], [5]]]
             return ground, eillposid
         # 重叠分区2
@@ -379,30 +379,30 @@ class Screen:
             for i in range(len(frag[1]))[2:]:
                 gmsh.model.occ.remove(frag[1][i], recursive=True)
 
-            ground = [[[16], [5, 0]],
+            ground = [[[16], [5, 0, 40, 44]],
                       [[18], [0]],
-                      [[20], [0, 1]],
+                      [[20], [0, 1, 47, 51]],
                       [[22], [1]],
-                      [[25], [1, 2]],
+                      [[25], [1, 2, 55, 59]],
                       [[24], [2]],
-                      [[23], [2, 3]],
+                      [[23], [2, 3, 57, 54]],
                       [[21], [3]],
-                      [[19], [3, 4]],
+                      [[19], [3, 4, 50, 46]],
                       [[17], [4]],
-                      [[15], [4, 5]],
+                      [[15], [4, 5, 43, 42]],
                       [[14], [5]]]
 
-            eillposid = [[[4], [5, 0]],
+            eillposid = [[[4], [5, 0, 3, 12]],
                          [[6], [0]],
-                         [[8], [0, 1]],
+                         [[8], [0, 1, 18, 24]],
                          [[10], [1]],
-                         [[12], [1, 2]],
+                         [[12], [1, 2, 30, 36]],
                          [[13], [2]],
-                         [[11], [2, 3]],
+                         [[11], [2, 3, 33, 27]],
                          [[9], [3]],
-                         [[7], [3, 4]],
+                         [[7], [3, 4, 21, 15]],
                          [[5], [4]],
-                         [[3], [4, 5]],
+                         [[3], [4, 5, 9, 6]],
                          [[2, 1], [5]]]
             return ground, eillposid
         else:
