+++ example/bem/backup/poisson_linear_dirichlet_2d_array.py
