+++ fealpy/bem/__init__.py
@@ -1,5 +1,6 @@
 from ..mesh import IntervalMesh, TriangleMesh, QuadrangleMesh
-from boundary_operator import BoundaryOperator
-from internal_operator import InternalOperator
-from potential_flux_integrator import PotentialFluxIntegrator
-from scalar_source_integrator import ScalarSourceIntegrator
+from .boundary_operator import BoundaryOperator
+from .internal_operator import InternalOperator
+from .potential_flux_integrator import PotentialFluxIntegrator
+from .scalar_source_integrator import ScalarSourceIntegrator
+from .boundary_condition import DirichletBC
